+++ android/documentation/changelog.md
@@ -1,5 +1,7 @@
 # Change Log
 <pre>
+v2.3.0 Added image rotation functionality.
+
 v2.2.1	[MOD-940] Building with 2.1.3.GA and open sourcing
 	
 v2.2	[MOD-447] Introduced "format" property. Default format is now JPEG, allowing for compression. Check out documentation to learn more.