+++ ios/documentation/changelog.md
@@ -1,5 +1,7 @@
 # Change Log
 <pre>
+v1.3.0 Added image rotation functionality.
+
 v1.2.0  [TIMOB-18092] Updated to build for 64bit
 
 v1.1.1	[MOD-940] Building with 2.1.3.GA and open sourcing